src/requirements.txt
ansible-core
ansible-lint
pylint
